backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14.0
Pillow>=10.0.0
django-cors-headers>=4.0.0

